+++ PCIT_core/internal/llvm_interface/src/IRBuilder.cpp
@@ -169,6 +169,9 @@ namespace pcit::llvmint{
 	}
 	
 	
+	auto IRBuilder::createBitCast(const Value& value, const Type& dst_type, evo::CStrProxy name) -> Value {
+		return Value(this->builder->CreateBitCast(value.native(), dst_type.native(), name.c_str()));
+	}	
 
 
 	//////////////////////////////////////////////////////////////////////